backend/requirements.txt
# FastAPI后端依赖（异步版本）
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
